changed place for color extensions
